Add files via upload
in this project i had loaded data using pandas, took in-sites about the data, presented relation and correlation between different features of this data. drooped the feature that is not useful to us. predicted about insurance claim using logistics regression model and decision tree model. treated the overfitting and compared the accuracy of the model